Add specs; update structure
There's a lot going on with this commit. I'm still unsure about the
structure yet. I may remove some of the nesting here.

I need to update the readme to demonstrate how to properly run this
again. I also need to find a way to test the !who action. I'm adding a
goofy !freak action as well, that I haven't been able to sort out the
best way to get in place yet.

Finally, I have a module that I'm using to act with constants for the
next meeting. I think I'm doing it wrong. I'll need to sort that out.